Update mock data and fix random title agains mock data

+++ sdk/com.ibm.sbt.core.test/src/test/java/com/ibm/sbt/services/client/connections/blogs/BlogPostUpdateDeleteTest.java
@@ -30,6 +30,7 @@ import org.junit.Test;
  */
 public class BlogPostUpdateDeleteTest extends BaseBlogServiceTest {
 
+	private boolean ignoreAfter = false;
 
 	@Before
 	public void initBlogPostTestData() {
@@ -37,15 +38,16 @@ public class BlogPostUpdateDeleteTest extends BaseBlogServiceTest {
 	}
 
 	@Test
-	public void CreateBlogPostTest() {
+	public void createBlogPostTest() {
 		try {
+			ignoreAfter = false;
 			BlogPost blogPost = new BlogPost(blogService, "");
 			blogPost.setTitle("Test Blog Post" + System.currentTimeMillis());
 			blogPost.setContent("Test Blog Post Content"+ System.currentTimeMillis());
 			BlogPost blogPostReturned = blogService.createBlogPost(blogPost, blog.getHandle());
 			assertNotNull(blogPostReturned.getTitle());
-			assertEquals(blogPost.getTitle(), blogPostReturned.getTitle());
-			assertEquals(blogPost.getContent(), blogPostReturned.getContent());
+			assertEquals(unRandomize(blogPost.getTitle()), unRandomize(blogPostReturned.getTitle()));
+			assertEquals(unRandomize(blogPost.getContent()), unRandomize(blogPostReturned.getContent()));
 			deleteBlogPost(blogPostReturned);
 		} catch (Exception e) {
 			fail("Error calling blogService.createBlogPost() caused by: "+e.getMessage());
@@ -54,27 +56,26 @@ public class BlogPostUpdateDeleteTest extends BaseBlogServiceTest {
 	}
 
 	@Test
-	public void GetBlogPost() {
+	public void getBlogPost() {
 		try {
 			BlogPost blogPostGot = blogService.getBlogPost(blog.getHandle(), blogPost.getPostUuid());
 
-			assertEquals(blogPost.getTitle(), blogPostGot.getTitle());
-			assertEquals(blogPost.getContent(), blogPostGot.getContent());
+			assertEquals(unRandomize(blogPost.getTitle()), unRandomize(blogPostGot.getTitle()));
+			assertEquals(unRandomize(blogPost.getContent()), unRandomize(blogPostGot.getContent()));
 		} catch (Exception e) {
 			fail("Error calling blogService.getBlog() caused by: "+e.getMessage());
 		}
 	}
 
 	@Test
-	public void UpdateBlogPost() {
-
+	public void updateBlogPost() {
 		try {
 			blogPost.setTitle("New title " + System.currentTimeMillis());
 			blogPost.setContent("New content " + System.currentTimeMillis());
 			blogService.updateBlogPost(blogPost, blogPost.getBlogHandle());
 			BlogPost postUpdated = blogService.getBlogPost(blogPost.getBlogHandle(), blogPost.getPostUuid());
-			assertEquals(blogPost.getTitle(), postUpdated.getTitle());
-			assertEquals(blogPost.getContent(), postUpdated.getContent());
+			assertEquals(unRandomize(blogPost.getTitle()), unRandomize(postUpdated.getTitle()));
+			assertEquals(unRandomize(blogPost.getContent()), unRandomize(postUpdated.getContent()));
 		} catch (Exception e) {
 			fail("Error calling blogService.updateBlog() caused by: "+e.getMessage());
 		}
@@ -83,20 +84,20 @@ public class BlogPostUpdateDeleteTest extends BaseBlogServiceTest {
 	@Test
 	public void deleteBlogPost() throws Exception {
 		try {
-			BlogPost createdBlogPost = createBlogPost();
-			BlogPost blogPostGot = blogService.getBlogPost(createdBlogPost.getBlogHandle(), createdBlogPost.getPostUuid());
-			assertEquals(createdBlogPost.getTitle(), blogPostGot.getTitle());
-			blogService.removeBlogPost(createdBlogPost.getPostUuid(), createdBlogPost.getBlogHandle());
-			blogService.getBlogPost(createdBlogPost.getBlogHandle(), createdBlogPost.getPostUuid());
+			ignoreAfter = true;
+			BlogPost blogPostGot = blogService.getBlogPost(blogPost.getBlogHandle(), blogPost.getPostUuid());
+			assertEquals(blogPost.getTitle(), blogPostGot.getTitle());
+			blogService.removeBlogPost(blogPost.getPostUuid(), blogPost.getBlogHandle());
+			blogService.getBlogPost(blogPost.getBlogHandle(), blogPost.getPostUuid());
 		}catch(Exception e) {
-				assertNotNull(e.getMessage());
+			assertNotNull(e.getMessage());
 		}
 	}
 
 	@After
 	public void deleteBlogPostTestDataOnExit() {
+		if (ignoreAfter) return;
 		try {
-
 			blogService.removeBlogPost(blogPost.getPostUuid(), blogPost.getBlogHandle());
 		} catch (BlogServiceException e) {
 			fail("Error calling blogService.removeBlogPost() caused by: "+e.getMessage());
